parsers/requirements.txt
# Newsletter Parser Requirements

# HTML Parsing
beautifulsoup4==4.12.2
lxml==5.1.0
html5lib==1.1

# Gmail API (if using Gmail integration)
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
